"Issue #69: Never overwrite p12 file
Refactor authentication main programs to use a shared `runAuthenticator` template function for improved modularity and code reuse."
